+++ src/rx-query.js
@@ -259,12 +259,6 @@ export class RxQuery {
                 sortArray.push(pushMe);
             });
             json.sort = sortArray;
-        } else {
-            // sort by primaryKey as default
-            // (always use _id because there is no primary-swap on json.sort)
-            json.sort = [{
-                _id: 'asc'
-            }];
         }
 
         if (options.limit) {